+++ apps/mobile/src/urql-shared/client.js
@@ -1,17 +1,16 @@
-import { createClient, fetchExchange, subscriptionExchange } from 'urql'
+import { createClient, fetchExchange } from 'urql'
 import { cacheExchange } from '@urql/exchange-graphcache'
 import { devtoolsExchange } from '@urql/devtools'
 import apiHost from 'util/apiHost'
 import { setSessionCookie } from 'util/session'
+import mobileSubscriptionExchange from './mobileSubscriptionExchange'
 import keys from './keys'
 import resolvers from './resolvers'
 import optimistic from './optimistic'
 import updates from './updates'
 import directives from './directives'
-import EventSource from 'react-native-sse'
-import { URL } from 'react-native-url-polyfill'
 
-const GRAPHQL_ENDPOINT_URL = `${apiHost}/noo/graphql`
+export const GRAPHQL_ENDPOINT_URL = `${apiHost}/noo/graphql`
 
 const cache = cacheExchange({
   keys,
@@ -26,55 +25,7 @@ const client = createClient({
     devtoolsExchange,
     cache,
     fetchExchange,
-    subscriptionExchange({
-      forwardSubscription: (operation) => {
-        const url = new URL(GRAPHQL_ENDPOINT_URL)
-
-        url.searchParams.append('query', operation.query)
-
-        if (operation.variables) {
-          url.searchParams.append(
-            'variables',
-            JSON.stringify(operation.variables)
-          )
-        }
-
-        return {
-          subscribe: (sink) => {
-            const eventSource = new EventSource(url.toString(), {
-              withCredentials: true, // This is required for cookies
-              credentials: 'include',
-              method: 'GET',
-              lineEndingCharacter: '\n'
-            })
-
-            eventSource.addEventListener('next', (event) => {
-              const data = JSON.parse(event.data)
-
-              sink.next(data)
-
-              if (eventSource.readyState === 2) {
-                sink.complete()
-              }
-            })
-
-            eventSource.addEventListener('error', (event) => {
-              if (event.type === 'error') {
-                console.error('Connection error:', event?.message)
-              } else if (event.type === 'exception') {
-                console.error('Error:', event?.message, event?.error)
-              }
-
-              sink.error(event)
-            })
-
-            return {
-              unsubscribe: () => eventSource.close()
-            }
-          }
-        }
-      }
-    })
+    mobileSubscriptionExchange
   ],
   fetch: async (...args) => {
     const response = await fetch(...args)
